commit from ibou

+++ src/app/app_core/notes/notes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, OnInit, Self } from '@angular/core';
+import { Injectable, OnInit, Self, Optional } from '@angular/core';
 import { AngularFirestore, AngularFirestoreCollection } from '@angular/fire/firestore';
 import { HttpClient } from '@angular/common/http';
 
@@ -14,6 +14,9 @@ import { Note } from './note.models';
 import {map} from 'rxjs/operators'
 import * as moment from 'moment'
 
+import { AdGroupService } from './ad-groupe.service'
+import {Ads} from './ads.service'
+
 
 @Injectable()
 export class NotesService implements OnInit {
@@ -26,12 +29,15 @@ export class NotesService implements OnInit {
   text_error_date = "Cette campagne a déjà commencé"
   error_end_date = "Date invalide ou campagne déjà terminée"
   private notesCollection: AngularFirestoreCollection<Note>;
+  
+ 
+  
+   
 
-
-  constructor(private afs: AngularFirestore, private auth: AuthService, private http: HttpClient) {
+  constructor(private auth: AuthService, private afs: AngularFirestore, private http: HttpClient,  private adGroupService: AdGroupService, private adsService : Ads) {
     this.auth.user.forEach(child => {
-        this.uid =child.uid
-        this.notesCollection = this.afs.collection('notes', (ref) => ref.where('owner', '==', child.uid)); 
+      this.uid =child.uid
+      this.notesCollection = this.afs.collection('notes', (ref) => ref.where('owner', '==', child.uid)); 
     })
   }
   ngOnInit() { 
@@ -68,7 +74,7 @@ export class NotesService implements OnInit {
          
           var startDate = moment(res['startDate'], "YYYYMMDD").fromNow()
           var endDate = moment(res['endDate'], "YYYYMMDD").fromNow()
-         this.createCampaign(res['id'], name, res['status_campaign'], res['startDate'], res['endDate'], res['startDateFrench'], res['endDateFrench'], res['servingStatus']).then(res=>{
+         this.createCampaign(res['id'], name, res['status_campaign'], res['startDate'], res['endDate'], res['startDateFrench'], res['endDateFrench'], res['servingStatus'], res['budgetId']).then(res=>{
             Swal.fire({
               title: 'Ajouter une nouvelle campagne',
               text: 'Campagne ajoutée avec succès',
@@ -162,6 +168,22 @@ export class NotesService implements OnInit {
       } */
     })
    
+  }
+  getListCampaign(uid: any) {
+   
+   
+   console.log(uid)
+ return this.afs.collection('notes', (ref) => ref.where('owner','==',`${uid}`)).snapshotChanges().pipe(
+      map((actions) => {
+        return actions.map((a) => {
+          const data = a.payload.doc.data();
+          return { id: a.payload.doc.id, ...data };
+        });
+      })
+    );
+
+    
+  
   }
   updateTargetLocation(id: string, campaign_id: string, name: string, location: any) {
  
@@ -380,7 +402,7 @@ export class NotesService implements OnInit {
   
 
 
-  prepareSaveCampaign(id: any, name: string, status: string, startDate: string, endDate: string, startDateFrench: string, endDateFrench: string ,servingStatus: string): Note {
+  prepareSaveCampaign(id: any, name: string, status: string, startDate: string, endDate: string, startDateFrench: string, endDateFrench: string ,servingStatus: string, budgetId: any): Note {
     const userDoc = this.afs.doc(`users/${this.uid}`);
       const newCampaign = {
       id_campagne: id,
@@ -390,7 +412,9 @@ export class NotesService implements OnInit {
         endDate: endDate,
         startDateFrench: startDateFrench,
         endDateFrench: endDateFrench,
-      servingStatus: servingStatus,
+        servingStatus: servingStatus,
+        account_money: 0, 
+      budgetId: budgetId,
       zones: [],
       createdAt: firebase.firestore.FieldValue.serverTimestamp(),
       createdBy: userDoc.ref,
@@ -425,8 +449,8 @@ export class NotesService implements OnInit {
       return this.getNote(id).update(data);
   }
 
-  async createCampaign(id_campagne: any, name: string, status: string, startDate: string,endDate: string, startDateFrench: string, endDateFrench: string, servingSatus: string) {
-    this.note = this.prepareSaveCampaign(id_campagne, name, status, startDate, endDate, startDateFrench, endDateFrench, servingSatus);
+  async createCampaign(id_campagne: any, name: string, status: string, startDate: string,endDate: string, startDateFrench: string, endDateFrench: string, servingSatus: string, budgetId: any) {
+    this.note = this.prepareSaveCampaign(id_campagne, name, status, startDate, endDate, startDateFrench, endDateFrench, servingSatus, budgetId);
     const docRef = await this.notesCollection.add(this.note);
   }
 
@@ -434,8 +458,109 @@ export class NotesService implements OnInit {
     return this.getNote(id).update(data);
   }
 
-  deleteNote(id: string) {
-    return this.getNote(id).delete();
+  deleteNote(id: string, ad_groups_list_id: any, ads_list_id: any) {
+    console.log(ad_groups_list_id.length)
+    console.log(ads_list_id.length)
+
+    console.log(ad_groups_list_id)
+    console.log(ads_list_id)
+    if (ad_groups_list_id.length == 0) {
+      console.log("pas de groupe , pas d'annonce")
+          return this.getNote(id).delete()
+
+    } else if (ad_groups_list_id.length == 1) {
+              console.log("un groupe d'annonce")
+      if (ads_list_id.length == 0) {
+        console.log("groupe d'annonce n'a aucune annonce")
+        this.deleteAdGroupList(ad_groups_list_id).then(res => {
+          if(res=="ok"){
+            return this.getNote(id).delete()
+
+          }
+        })
+      } else {
+         console.log(`ads list len ${ads_list_id.length}`)
+        console.log("groupe d'annonce a plusieurs annonces")
+         this.deleteAdList(ads_list_id).then(res => {
+          this.deleteAdGroupList(ad_groups_list_id).then(res => {
+          return this.getNote(id).delete()
+         
+          })
+        })
+      }
+    
+      
+     
+    } else if (ad_groups_list_id.length > 1) {
+      console.log('plusieurs groupes')
+      if (ads_list_id.length == 0) {
+         this.deleteAdGroupList(ad_groups_list_id).then(res => {
+          return this.getNote(id).delete()
+         
+          })
+      } else {
+        
+        this.deleteAdList(ads_list_id).then(res => {
+             this.deleteAdGroupList(ad_groups_list_id).then(res => {
+             return this.getNote(id).delete()
+            
+             })
+           })
+      }
+    } 
+  }
+
+  deleteAdList(ads_list_id: any): Promise<any> {
+    return new Promise(resolve => {
+      if (ads_list_id.length == 1) {
+          
+        this.adsService.deleteAd(ads_list_id[0]).then(res => {
+          
+          resolve('ok')
+        })
+        
+          
+        }else {
+         let i = 0
+         console.log('groupe')
+         console.log(ads_list_id)
+         for (let i = 0; i < ads_list_id.length; i++){
+             console.log('removing')
+            this.adsService.deleteAd(ads_list_id[i]).then(res => {
+          
+          resolve('ok')
+        })
+         }
+       
+      }
+      
+    })
+  }
+   deleteAdGroupList(ad_groups_list_id: any): Promise<any> {
+     return new Promise(resolve => {
+       if (ad_groups_list_id.length == 1) {
+         this.adGroupService.deleteAdGroup(ad_groups_list_id[0]).then(res => {
+           
+             
+           resolve('ok')
+         })
+          
+        }else {
+         let i = 0
+         console.log('groupe')
+         console.log(ad_groups_list_id)
+         for (let i = 0; i < ad_groups_list_id.length; i++){
+             console.log('removing')
+            this.adGroupService.deleteAdGroup(ad_groups_list_id[i]).then(res => {
+              resolve('ok')
+           
+            })
+         }
+       
+      }
+       
+      
+    })
   }
   
 }
